Changed collision object ids to int, added constructors with id.

+++ core/src/main/java/org/mini2Dx/core/engine/Sizeable.java
@@ -20,7 +20,7 @@ public interface Sizeable {
 	 * 
 	 * @return
 	 */
-	public long getId();
+	public int getId();
 
 	/**
 	 * Returns the width of this object
